refactoring App to a new root level components namespace
