+++ src/components/App/translations/en.json
